dapp-update: moved HTS contracts API folders to /api/hts-interactions
