+++ system-contract-dapp-playground/__tests__/hedera/hts-interactions/token-management-contract/index.test.ts
@@ -24,7 +24,7 @@ import {
   manageTokenDeduction,
   manageTokenInfomation,
   manageTokenPermission,
-} from '@/api/hedera/tokenManagement-interactions';
+} from '@/api/hedera/hts-interactions/tokenManagement-interactions';
 import {
   CommonKeyObject,
   IHederaTokenServiceExpiry,
